% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/n_subject.R
\name{avg_duration}
\alias{avg_duration}
\title{Average number of events}
\usage{
avg_duration(id, group, dur, par = NULL)
}
\arguments{
\item{id}{a character vector of subject id}

\item{group}{a factor vector of group name}

\item{dur}{a numeric vector of AE duration}

\item{par}{a character vector of parameter name}
}
\description{
Average number of events
}
\examples{
library(r2rtf)
r2rtf_adae$TRTA <- factor(r2rtf_adae$TRTA)
avg_event(r2rtf_adae$USUBJID, r2rtf_adae$TRTA)
avg_event(r2rtf_adae$USUBJID, r2rtf_adae$TRTA, r2rtf_adae$AEDECOD)

}
