% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/collect_ae_listing.R
\name{collect_ae_listing}
\alias{collect_ae_listing}
\title{Add inference information for AE listing analysis}
\usage{
collect_ae_listing(
  outdata,
  display = c("SEX", "RACE", "AGE", "ASTDY", "AESEV", "AESER", "AEREL", "AEACN", "AEOUT",
    "SITEID", "ADURN", "ADURU")
)
}
\arguments{
\item{outdata}{a \code{outdata} object created by \code{prepare_ae_specific}}

\item{display}{a vector with name of variable used to display on AE listing.}
}
\description{
Add inference information for AE listing analysis
}
\examples{
library(metalite)
meta <- meta_ae_dummy()

lapply(prepare_ae_specific(meta, "apat", "wk12", "rel") |>
collect_ae_listing(), head, 10)

}
