% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extend_ae_specific.R
\name{extend_ae_specific_events}
\alias{extend_ae_specific_events}
\title{Add average number of events information for AE specific analysis}
\usage{
extend_ae_specific_events(outdata)
}
\arguments{
\item{outdata}{An \code{outdata} object created by \code{\link[=prepare_ae_specific]{prepare_ae_specific()}}.}
}
\value{
A list of analysis raw datasets.
}
\description{
Add average number of events information for AE specific analysis
}
\examples{
meta <- metalite::meta_example()
tbl <- prepare_ae_specific(meta,
  population = "apat",
  observation = "wk12",
  parameter = "rel"
) |>
  extend_ae_specific_events() |>
  format_ae_specific(display = c("n", "prop", "events_avg"))
head(tbl$tbl)
}
