% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extend_ae_specific.R
\name{extend_ae_specific_subgroup}
\alias{extend_ae_specific_subgroup}
\title{Add subgroup analysis in AE specific analysis}
\usage{
extend_ae_specific_subgroup(outdata, subgroup_var)
}
\arguments{
\item{outdata}{An \code{outdata} object created by \code{\link[=prepare_ae_specific]{prepare_ae_specific()}}.}

\item{subgroup_var}{a character string for subgroup variable name}
}
\value{
A list of analysis raw datasets.
}
\description{
Add subgroup analysis in AE specific analysis
}
\examples{
meta <- metalite::meta_example()
tbl <- prepare_ae_specific(meta,
  population = "apat",
  observation = "wk12",
  parameter = "rel"
) |>
  extend_ae_specific_subgroup(subgroup_var = "SEX")
}
