% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fmt.R
\name{fmt_ci}
\alias{fmt_ci}
\title{Format confidence interval}
\usage{
fmt_ci(lower, upper, digits = 2, width = 3 + digits)
}
\arguments{
\item{lower}{A numeric value of lower value of CI.}

\item{upper}{A numeric value of upper value of CI.}

\item{digits}{Digits of each column, i.e., format as (x.x, x.x).}

\item{width}{Width of each column.}
}
\value{
To be added.
}
\description{
Format confidence interval
}
\examples{
# To be added
}
