% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fmt.R
\name{fmt_pct}
\alias{fmt_pct}
\title{Format percentage}
\usage{
fmt_pct(x, digits = 1, pre = "(", post = ")")
}
\arguments{
\item{x}{a numeric vector}

\item{digits}{number of digits}

\item{pre}{text before the number}

\item{post}{text after the number}
}
\description{
Format percentage
}
\examples{
fmt_pct(c(1, 1.52, 0.3, 100))

}
