% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fmt.R
\name{fmt_pval}
\alias{fmt_pval}
\title{Format P-Value}
\usage{
fmt_pval(p, digits = 3, width = 3 + digits)
}
\arguments{
\item{p}{a numeric vector of p-values}

\item{digits}{digits of each column, i.e. format as x.xxx}

\item{width}{width of each column}
}
\description{
Format P-Value
}
