% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meta_ae_dummy.R
\name{meta_ae_listing_dummy}
\alias{meta_ae_listing_dummy}
\title{Create a dummy \code{meta_adam} object for AE listing}
\usage{
meta_ae_listing_dummy()
}
\description{
This function is only for illustration purpose.
\code{r2rtf} and \code{magrittr} packages are required.
}
