% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{metalite.ae_adex}
\alias{metalite.ae_adex}
\title{ADEX dataset.}
\format{
A data frame with 591 rows and 41 variables:
}
\source{
\url{https://github.com/phuse-org/phuse-scripts/tree/master/data/sdtm/cdiscpilot01}
}
\usage{
metalite.ae_adex
}
\description{
A dataset containing exposure detail.
}
\keyword{datasets}
