% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{metalite_ae_adexsum}
\alias{metalite_ae_adexsum}
\title{ADEXSUM dataset}
\format{
A data frame with 254 rows and 30 variables.
}
\source{
\url{https://github.com/phuse-org/phuse-scripts/tree/master/data/sdtm/cdiscpilot01}
}
\usage{
metalite_ae_adexsum
}
\value{
An analysis data frame.
}
\description{
A dataset containing exposure details in Basic Data Structure (BDS).
}
\keyword{datasets}
