% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_ae_exp_adj.R
\name{prepare_ae_exp_adj}
\alias{prepare_ae_exp_adj}
\title{Prepare datasets for AE specific analysis}
\usage{
prepare_ae_exp_adj(
  meta,
  population = "apat",
  observation = "wk12",
  parameter = "any;rel;ser",
  adj_unit = c("year", "month", "week", "day"),
  ...
)
}
\value{
A list of analysis raw datasets.
}
\description{
Prepare datasets for AE specific analysis
}
