% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format_ae_listing.R
\name{propercase}
\alias{propercase}
\title{Function to propcase character variable}
\usage{
propercase(x)
}
\arguments{
\item{x}{a character vector to propcase}
}
\description{
Function to propcase character variable
}
\examples{

propercase("hello world")

}
