% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tlf_ae_exp_adj.R
\name{tlf_ae_exp_adj}
\alias{tlf_ae_exp_adj}
\title{Exposure-adjusted AE summary table}
\usage{
tlf_ae_exp_adj(
  outdata,
  source,
  analysis,
  col_rel_width = NULL,
  text_font_size = 9,
  orientation = "portrait",
  title = c("analysis", "observation", "population"),
  footnotes = NULL,
  path_outdata = NULL,
  path_outtable = NULL
)
}
\arguments{
\item{outdata}{An \code{outdata} object created by \code{\link[=prepare_ae_specific]{prepare_ae_specific()}}.}

\item{source}{A character value of the data source.}

\item{analysis}{One of analysis name existing at \code{outdata$meta$analysis}}

\item{col_rel_width}{Column relative width in a vector e.g. c(2,1,1) refers to 2:1:1.
Default is NULL for equal column width.}

\item{text_font_size}{Text font size.  To vary text font size by column, use
numeric vector with length of vector equal to number of columns
displayed e.g. c(9,20,40).}

\item{orientation}{Orientation in 'portrait' or 'landscape'.}

\item{title}{Term "analysis", "observation"and "population") for collecting
title from metadata or a character vector of table titles.}

\item{footnotes}{A character vector of table footnotes.}

\item{path_outdata}{A character string of the outdata path.}

\item{path_outtable}{A character string of the outtable path.}
}
\value{
RTF file and source dataset for exposure-adjusted AE summary table.
}
\description{
Exposure-adjusted AE summary table
}
\examples{
meta <- metalite::meta_example()
meta <- meta |>
  metalite::define_plan(
    plan = metalite::add_plan(meta$plan,
      analysis = "ae_exp_adj", population = "apat",
      observation = "wk12", parameter = "any"
    )
  ) |>
  metalite::define_analysis(
    name = "ae_exp_adj",
    label = "Exposure Adjusted Incident Rate",
    title = "Exposure-Adjusted Adverse Event Summary"
  ) |>
  metalite::meta_build()
outdata <- meta |>
  prepare_ae_summary(
    population = "apat",
    observation = "wk12",
    parameter = "any;rel;ser"
  ) |>
  extend_ae_summary_eaer(adj_unit = "month")
outdata |>
  format_ae_exp_adj() |>
  tlf_ae_exp_adj(
    source = "Source:  [CDISCpilot: adam-adsl; adae]",
    analysis = "ae_exp_adj",
    path_outdata = tempfile(fileext = ".Rdata"),
    path_outtable = tempfile(fileext = ".rtf")
  )
}
