% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/to_sentence.R
\name{to_sentence}
\alias{to_sentence}
\title{To Sentence Style}
\usage{
to_sentence(x)
}
\arguments{
\item{x}{a character vector}
}
\description{
To Sentence Style
}
\examples{
to_sentence("this is A Example")
}
