% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extend_ae_specific.R
\name{extend_ae_specific_duration}
\alias{extend_ae_specific_duration}
\title{Add average duration information for AE specific analysis}
\usage{
extend_ae_specific_duration(outdata, duration_var, duration_unit = "Day")
}
\arguments{
\item{outdata}{a \code{outdata} object created by \code{prepare_ae_specific}}

\item{duration_var}{a character value of variable name for AE duration.}

\item{duration_unit}{a character value of AE duration unit.}
}
\description{
Add average duration information for AE specific analysis
}
