% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extend_ae_specific.R
\name{extend_ae_specific_events}
\alias{extend_ae_specific_events}
\title{Add average number of events information for AE specific analysis}
\usage{
extend_ae_specific_events(outdata)
}
\arguments{
\item{outdata}{A \code{outdata} object created by \code{\link[=prepare_ae_specific]{prepare_ae_specific()}}}
}
\value{
To be added.
}
\description{
Add average number of events information for AE specific analysis
}
\examples{
# To be added
}
