% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extend_ae_specific.R
\name{extend_ae_specific_inference}
\alias{extend_ae_specific_inference}
\title{Add inference information for AE specific analysis}
\usage{
extend_ae_specific_inference(outdata, ci = 0.95)
}
\arguments{
\item{outdata}{a \code{outdata} object created by \code{prepare_ae_specific}}

\item{ci}{a numeric value for the percentile of confidence interval.}
}
\description{
Add inference information for AE specific analysis
}
