% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fmt.R
\name{fmt_ci}
\alias{fmt_ci}
\title{Format Confidence Interval}
\usage{
fmt_ci(lower, upper, digits = 2, width = 3 + digits)
}
\arguments{
\item{lower}{a numeric value of lower value of CI}

\item{upper}{a numeric value of lower value of CI}

\item{digits}{digits of each column, i.e. format as (x.x, x.x)}

\item{width}{width of each column}
}
\description{
Format Confidence Interval
}
