% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fmt.R
\name{fmt_pval}
\alias{fmt_pval}
\title{Format p-value}
\usage{
fmt_pval(p, digits = 3, width = 3 + digits)
}
\arguments{
\item{p}{A numeric vector of p-values.}

\item{digits}{Digits of each column, i.e., format as x.xxx.}

\item{width}{Width of each column.}
}
\value{
To be added.
}
\description{
Format p-value
}
\examples{
# To be added
}
