% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format_ae_listing.R
\name{format_ae_listing}
\alias{format_ae_listing}
\title{Format AE listing analysis}
\usage{
format_ae_listing(outdata, display_subline = FALSE)
}
\arguments{
\item{outdata}{a \code{outdata} object created by \code{prepare_ae_specific}}

\item{display_subline}{whether display subline or not.}
}
\description{
Format AE listing analysis
}
\examples{
library(metalite)
meta <- meta_ae_dummy()

lapply(prepare_ae_specific(meta, "apat", "wk12", "rel") |>
collect_ae_listing() |>
format_ae_listing(), head, 100)

}
