% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meta_ae_example.R
\name{meta_ae_example}
\alias{meta_ae_example}
\title{Create an example \code{meta_adam} object}
\usage{
meta_ae_example()
}
\value{
A metadata object.
}
\description{
This function is only for illustration purpose.
r2rtf is required.
}
\examples{
meta <- meta_ae_example()
}
