% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package.R
\docType{package}
\name{metalite.ae-package}
\alias{metalite.ae}
\alias{metalite.ae-package}
\title{metalite.ae: AE Analysis Using `metalite`}
\description{
An R package for adverse events analysis using `metalite` data structure.
}
\author{
\strong{Maintainer}: Nan Xiao \email{nan.xiao1@merck.com}

Authors:
\itemize{
  \item Yilong Zhang
  \item Xuan Deng
  \item Ruchitbhai Patel
  \item Madhusudhan Ginnaram
  \item Sarad Nepal
  \item Yujie Zhao
}

Other contributors:
\itemize{
  \item Merck Sharp & Dohme Corp [copyright holder]
}

}
\keyword{internal}
