% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package.R
\docType{package}
\name{metalite.ae-package}
\alias{metalite.ae}
\alias{metalite.ae-package}
\title{metalite.ae: Adverse Events Analysis Using 'metalite'}
\description{
\if{html}{\figure{logo.png}{options: style='float: right' alt='logo' width='120'}}

An R package to analyze adverse events in clinical trials using the 'metalite' data structure. The R package simplified the workflow to create production ready table, listing and figures discussed in the Chpater 7 and Chapter 8 of the R for Clinical Study Reports and Submission book (https://r4csr.org/).
}
\seealso{
Useful links:
\itemize{
  \item \url{https://merck.github.io/metalite.ae/}
  \item \url{https://github.com/Merck/metalite.ae}
  \item Report bugs at \url{https://github.com/Merck/metalite.ae/issues}
}

}
\author{
\strong{Maintainer}: Yujie Zhao \email{yujie.zhao@merck.com}

Authors:
\itemize{
  \item Yilong Zhang
  \item Xuan Deng
  \item Ruchitbhai Patel
  \item Madhusudhan Ginnaram
  \item Sarad Nepal
  \item Brian Lang
  \item Benjamin Wang
  \item Nan Xiao
}

Other contributors:
\itemize{
  \item Merck Sharp & Dohme Corp [copyright holder]
}

}
\keyword{internal}
