% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package.R
\docType{package}
\name{metalite.ae-package}
\alias{metalite.ae}
\alias{metalite.ae-package}
\title{metalite.ae: Adverse Events Analysis Using 'metalite'}
\description{
Adverse events analysis using the 'metalite' data structure.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://merck.github.io/metalite.ae/}
  \item \url{https://github.com/Merck/metalite.ae}
  \item Report bugs at \url{https://github.com/Merck/metalite.ae/issues}
}

}
\author{
\strong{Maintainer}: Nan Xiao \email{nan.xiao1@merck.com}

Authors:
\itemize{
  \item Yilong Zhang
  \item Xuan Deng
  \item Ruchitbhai Patel
  \item Madhusudhan Ginnaram
  \item Sarad Nepal
  \item Yujie Zhao
}

Other contributors:
\itemize{
  \item Merck Sharp & Dohme Corp [copyright holder]
}

}
\keyword{internal}
