% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package.R
\docType{package}
\name{metalite.ae-package}
\alias{metalite.ae}
\alias{metalite.ae-package}
\title{metalite.ae: Adverse Events Analysis Using 'metalite'}
\description{
\if{html}{\figure{logo.png}{options: style='float: right' alt='logo' width='120'}}

Analyzes adverse events in clinical trials using the 'metalite' data structure. The package simplifies the workflow to create production-ready tables, listings, and figures discussed in the adverse events analysis chapters of "R for Clinical Study Reports and Submission" by Zhang et al. (2022) \url{https://r4csr.org/}.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://merck.github.io/metalite.ae/}
  \item \url{https://github.com/Merck/metalite.ae}
  \item Report bugs at \url{https://github.com/Merck/metalite.ae/issues}
}

}
\author{
\strong{Maintainer}: Yujie Zhao \email{yujie.zhao@merck.com}

Authors:
\itemize{
  \item Yilong Zhang
  \item Benjamin Wang
  \item Nan Xiao
  \item Sarad Nepal
  \item Madhusudhan Ginnaram
  \item Ruchitbhai Patel
  \item Brian Lang
  \item Xuan Deng
  \item Hiroaki Fukuda
  \item Bing Liu
}

Other contributors:
\itemize{
  \item Venkatesh Burla [contributor]
  \item Merck Sharp & Dohme Corp [copyright holder]
}

}
\keyword{internal}
