% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{metalite.ae_adex}
\alias{metalite.ae_adex}
\title{ADEX dataset.}
\format{
A data frame with 280901 rows and 4 variables:
\describe{
\item{ASTDT}{Analysis Start Date.}
\item{AENDT}{Analysis End Date.}
\item{ASTDTM}{Analysis Start Date Time.}
\item{AENDTM}{Analysis End Date Time.}
\item{ASTDY}{Analysis Study Day.}
\item{AENDY}{Exposure Duration.}
\item{EXDURDD}{Analysis Start Date.}
\item{EXNUMDOS}{Number of Daily Doses.}
}
}
\source{
\url{https://github.com/phuse-org/phuse-scripts/tree/master
}
\usage{
metalite.ae_adex
}
\description{
A dataset containing exposure detail.
}
\keyword{datasets}
