% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/n_subject.R
\name{n_subject}
\alias{n_subject}
\title{Count number of unique subjects}
\usage{
n_subject(id, group, par = NULL)
}
\arguments{
\item{id}{a character vector of subject id}

\item{group}{a factor vector of group name}

\item{par}{a character vector of parameter name}
}
\description{
Count number of unique subjects
}
\examples{
library(r2rtf)
r2rtf_adae$TRTA <- factor(r2rtf_adae$TRTA)
metalite.ae:::n_subject(r2rtf_adae$USUBJID, r2rtf_adae$TRTA)
metalite.ae:::n_subject(r2rtf_adae$USUBJID, r2rtf_adae$TRTA, r2rtf_adae$AEDECOD)
}
