% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_ae_exp_adj.R
\name{prepare_ae_exp_adj}
\alias{prepare_ae_exp_adj}
\title{Prepare datasets for AE specific analysis}
\usage{
prepare_ae_exp_adj(
  meta,
  population = "apat",
  observation = "wk12",
  parameter = "any;rel;ser",
  duration_var = "TRTDUR",
  adj_unit = c("year", "month", "week", "day")
)
}
\arguments{
\item{meta}{A metadata object created by metalite.}

\item{population}{A character value of population term name.
The term name is used as key to link information.}

\item{observation}{A character value of observation term name.
The term name is used as key to link information.}

\item{parameter}{A character value of parameter term name.
The term name is used as key to link information.}

\item{duration_var}{A character value of duration variable name.}

\item{adj_unit}{A character value of exposure adjusted unit.}
}
\value{
A list of analysis raw datasets.
}
\description{
Prepare datasets for AE specific analysis
}
