% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_ae_listing.R
\name{prepare_ae_listing}
\alias{prepare_ae_listing}
\title{Prepare datasets for AE specific analysis}
\usage{
prepare_ae_listing(meta, population, observation, parameter)
}
\arguments{
\item{meta}{a meta data created by \code{metalite}.}

\item{population}{a character value of population term name.
The term name is used as key to link information.}

\item{observation}{a character value of observation term name.
The term name is used as key to link information.}

\item{parameter}{a character value of parameter term name.
The term name is used as key to link information.}
}
\description{
Prepare datasets for AE specific analysis
}
\examples{
\dontrun{
}
}
