% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_ae_listing.R
\name{prepare_ae_listing}
\alias{prepare_ae_listing}
\title{Prepare datasets for AE specific analysis}
\usage{
prepare_ae_listing(meta, analysis, population, observation, parameter)
}
\arguments{
\item{meta}{a meta data created by \code{metalite}.}

\item{analysis}{analysis name from \code{meta}}

\item{population}{a character value of population term name.
The term name is used as key to link information.}

\item{observation}{a character value of observation term name.
The term name is used as key to link information.}

\item{parameter}{a character value of parameter term name.
The term name is used as key to link information.}
}
\value{
To be added.
}
\description{
Prepare datasets for AE specific analysis
}
\examples{
meta <- meta_ae_listing_example()
lapply(prepare_ae_listing(meta, "ae_listing", "apat", "wk12", "ser"), head, 10)
}
