% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_ae_specific_subgroup.R
\name{prepare_ae_specific_subgroup}
\alias{prepare_ae_specific_subgroup}
\title{Prepare datasets for AE specific analysis}
\usage{
prepare_ae_specific_subgroup(
  meta,
  population,
  observation,
  parameter,
  subgroup_var,
  subgroup_header = 1,
  components = c("soc", "par"),
  display_total = TRUE,
  display_subgroup_total = TRUE
)
}
\arguments{
\item{meta}{a meta data created by \code{metalite}.}

\item{population}{a character value of population term name.
The term name is used as key to link information.}

\item{observation}{a character value of observation term name.
The term name is used as key to link information.}

\item{parameter}{a character value of parameter term name.
The term name is used as key to link information.}

\item{subgroup_var}{a character value of subgroup variable name in observation data saved in \code{meta$data_observation}.}

\item{subgroup_header}{a integer value of column header for subgroup.}

\item{components}{a character vector of components name.}
}
\description{
Prepare datasets for AE specific analysis
}
\examples{
meta <- meta_ae_dummy()
prepare_ae_specific_subgroup(meta, "apat", "wk12", "rel", subgroup_var = "RACE")$data
}
