% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_ae_specific_subgroup.R
\name{prepare_ae_specific_subgroup}
\alias{prepare_ae_specific_subgroup}
\title{Prepare datasets for AE specific analysis}
\usage{
prepare_ae_specific_subgroup(
  meta,
  population,
  observation,
  parameter,
  subgroup_var,
  subgroup_header = 1,
  components = c("soc", "par"),
  display_total = TRUE,
  display_subgroup_total = TRUE
)
}
\arguments{
\item{meta}{A metadata object created by metalite.}

\item{population}{A character value of population term name.
The term name is used as key to link information.}

\item{observation}{A character value of observation term name.
The term name is used as key to link information.}

\item{parameter}{A character value of parameter term name.
The term name is used as key to link information.}

\item{subgroup_var}{A character value of subgroup variable name in
observation data saved in \code{meta$data_observation}.}

\item{subgroup_header}{An integer value of column header for subgroup.}

\item{components}{A character vector of components name.}

\item{display_total}{test}

\item{display_subgroup_total}{test}
}
\value{
A list of analysis raw datasets.
}
\description{
Prepare datasets for AE specific analysis
}
\examples{
# meta <- meta_ae_example()
# prepare_ae_specific_subgroup(
#   meta,
#   population = "apat",
#   observation = "wk12",
#   parameter = "rel",
#   subgroup_var = "RACE"
# )$data
}
