% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_ae_summary.R
\name{prepare_ae_summary}
\alias{prepare_ae_summary}
\title{Prepare datasets for AE specific analysis}
\usage{
prepare_ae_summary(meta, population, observation, parameter, ...)
}
\arguments{
\item{meta}{a meta data created by \code{metalite}.}

\item{population}{a character value of population term name.
The term name is used as key to link information.}

\item{observation}{a character value of observation term name.
The term name is used as key to link information.}

\item{parameter}{a character value of parameter term name.
The term name is used as key to link information.}

\item{...}{additional parameters transfer to \code{prepare_ae_specific}}
}
\description{
Prepare datasets for AE specific analysis
}
\examples{
meta <- meta_ae_dummy()
prepare_ae_summary(meta, "apat", "wk12", "any;rel;ser")
}
