% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_ae_summary.R
\name{prepare_ae_summary}
\alias{prepare_ae_summary}
\title{Prepare datasets for AE specific analysis}
\usage{
prepare_ae_summary(meta, population, observation, parameter, ...)
}
\arguments{
\item{meta}{A metadata object created by metalite.}

\item{population}{A character value of population term name.
The term name is used as key to link information.}

\item{observation}{A character value of observation term name.
The term name is used as key to link information.}

\item{parameter}{A character value of parameter term name.
The term name is used as key to link information.}

\item{...}{Additional arguments passed to \code{\link[=prepare_ae_specific]{prepare_ae_specific()}}.}
}
\value{
An objects contain a list of analysis raw datasets.
}
\description{
Prepare datasets for AE specific analysis
}
\examples{
meta <- meta_ae_example()
prepare_ae_summary(
  meta,
  population = "apat",
  observation = "wk12",
  parameter = "any;rel;ser"
)
}
