% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tlf_ae_specific.R
\name{tlf_ae_specific}
\alias{tlf_ae_specific}
\title{Specific adverse events table}
\usage{
tlf_ae_specific(
  outdata,
  medra_version,
  source,
  col_rel_width = NULL,
  text_font_size = 9,
  orientation = "portrait",
  footnotes = NULL,
  title = NULL,
  path_outdata = NULL,
  path_outtable = NULL
)
}
\arguments{
\item{outdata}{a outdata list created from \code{prepare_ae_specific}}

\item{medra_version}{a character value of the MedDRA Version for this dataset}

\item{source}{a character value of the data source}

\item{col_rel_width}{Column relative width in a vector e.g. c(2,1,1) refers to 2:1:1.
Default is NULL for equal column width.}

\item{text_font_size}{Text font size.  To vary text font size by column, use
numeric vector with length of vector equal to number of columns
displayed e.g. c(9,20,40).}

\item{orientation}{Orientation in 'portrait' or 'landscape'.}

\item{footnotes}{a character vector of table footnotes}

\item{title}{a character vector of table titles}

\item{path_outdata}{a character string of the outdata path}

\item{path_outtable}{a character string of the outtable path}
}
\value{
To be added.
}
\description{
Specific adverse events table
}
\examples{
meta <- meta_ae_example()

meta |>
  prepare_ae_specific(
    population = "apat",
    observation = "wk12",
    parameter = "rel"
  ) |>
  format_ae_specific() |>
  tlf_ae_specific(
    source = "Source:  [CDISCpilot: adam-adsl; adae]",
    medra_version = "24.0",
    path_outdata = tempfile(fileext = ".Rdata"),
    path_outtable = tempfile(fileext = ".rtf")
  )
}
